+++ packages/core/test/index.js
@@ -4,13 +4,7 @@ import renderer from 'react-test-renderer'
 import { render, fireEvent, cleanup, act } from '@testing-library/react'
 import { matchers } from 'jest-emotion'
 import mockConsole from 'jest-mock-console'
-import {
-  jsx,
-  Context,
-  useThemeUI,
-  merge,
-  ThemeProvider,
-} from '../src'
+import { jsx, Context, useThemeUI, merge, ThemeProvider } from '../src'
 
 afterEach(cleanup)
 
@@ -21,7 +15,7 @@ const renderJSON = el => renderer.create(el).toJSON()
 describe('ThemeProvider', () => {
   test('renders', () => {
     const json = renderJSON(
-      <ThemeProvider>
+      <ThemeProvider theme={{}}>
         <h1>Hello</h1>
       </ThemeProvider>
     )
@@ -35,7 +29,7 @@ describe('ThemeProvider', () => {
         value={{
           emotionVersion: '9.0.0',
         }}>
-        <ThemeProvider>Conflicting versions</ThemeProvider>
+        <ThemeProvider theme={{}}>Conflicting versions</ThemeProvider>
       </Context.Provider>
     )
     expect(console.warn).toBeCalled()
@@ -104,20 +98,21 @@ describe('ThemeProvider', () => {
       cards: {
         default: {
           border: t => `1px solid ${t.colors.primary}`,
-        }
-      }
+        },
+      },
     }
     const json = renderJSON(
-      jsx(ThemeProvider, { theme },
+      jsx(
+        ThemeProvider,
+        { theme },
         jsx('div', {
           sx: {
             variant: 'cards.default',
-          }
+          },
         })
       )
     )
     expect(json).toHaveStyleRule('border', '1px solid tomato')
-
   })
 })
 
@@ -151,13 +146,15 @@ describe('jsx', () => {
 
   test('css prop accepts functions', () => {
     const json = renderJSON(
-      jsx(ThemeProvider, {
-        theme: {
-          colors: {
-            primary: 'tomato',
-          }
-        }
-      },
+      jsx(
+        ThemeProvider,
+        {
+          theme: {
+            colors: {
+              primary: 'tomato',
+            },
+          },
+        },
         jsx('div', {
           css: t => ({
             color: t.colors.primary,
@@ -338,7 +335,7 @@ describe('useThemeUI', () => {
         theme={{
           colors: {
             text: 'tomato',
-          }
+          },
         }}>
         <GetContext />
       </ThemeProvider>
@@ -347,4 +344,3 @@ describe('useThemeUI', () => {
     expect(context.theme.colors.text).toBe('tomato')
   })
 })
-